test_default_policy_is_tablet_aware: Better tablet info filling
This commit switches from using double loop with 100 executions in each
iter, to sending statement everywhere for each tablet.
Why?
- It is easier to understand, so no more wall of text
- It is faster (~1.2 seconds faster).
- It is more resilient, because we have 100% confidence that
  each statement reached non-replica.
One thing we lose is that test no longer checks receiving feedbacks
when using DefaultPolicy, but this is not really useful. The important
part is checking that we receive no feedbacks when we have info. The
only way for the removed part to fail (apart from migrations and other
stuff Scylla inflicts on us) is if driver can magically predict tablet
info. I'd be happy if it gets magical powers, so I'm not worried about
the removed part.