ser/deser tests: Make sure structs are used
In Rust 1.89, those structs started to trigger dead_code warning with
information that they are never constructed.
In row_tests / value_tests I don't want to allow/expect dead_code,
because part of the reason for existence of those structs is to verify
generated impls for those structs.
The solution I chose is to write a never-called function that calles the
relevant deserialize impl. This seems to work.
Curiously, TestRowByName in hygiene tests doesn't trigger the warning. I
am unable to tell why. I tried to replace derives with their generated
impls, but I didn't notice any difference that could have caused this.