Preparing for 2024 edition: automatic changes
How this commit was created:
I executed `cargo fix --edition`. After that I looked through the
changes, and reverted the ones I thought were unnecessary.
What changes I reverted?
- Rust 2024 added new expression types matched by `expr` fragment
specifier. For backwards compatibility `expr_2021` exists as well. cargo
moves the code to use it, but for our macros `expr` is still fine - they
are only internal after all.
- In Rust 2024, lifetime of temporary expression in `if let Pattern =
Expression` changed. Now it is dropped before entering `else` branch.
Cargo switches the code to use `match` instead of `if let` for
compatibility. I reviewed affected places in our code and concluded that
we can keep using `if let` everywhere.

That means that the only commited changes are related to RPIT lifetime
capture rules. Cargo introduced `use` bounds in several places. Most of
them are not important since they are not pub so we can change them
later. Two places were pub, and cargo correctly preserved current
semantics (which are also reasonable imo).
The pub places:
- `ColumnSpecs::iter`: `use<'slice, 'spec>` was introduced. We don't
need or want to bind to lifetime of a self reference, only to lifetimes
of held slice, so it makes sense.
- `Sharder::iter_source_ports_for_shard`. `use<>` was introduced by
Cargo. This is correct - the returned iterator does not depend on self.
self is only needed to get the bounds of the port range.

There is one thing that automatic migration doesn't handle the "outlives
trick":
https://doc.rust-lang.org/edition-guide/rust-2024/rpit-lifetime-capture.html#migrating-away-from-the-outlives-trick
I don't understand it completely unfortunately. We should understand it,
and if needed migrate from it, before migrating to 2024 edition.
We have a method like this for example:
```
    pub(crate) fn iter_working_connections_per_node(
        &self,
    ) -> Result<
        impl Iterator<Item = (Uuid, impl Iterator<Item =
Arc<Connection>> + use<>)> + '_,
        ConnectionPoolError,
    >
```
Is it an example of this trick? Should we change this `+ '_` to
`use<'_>`?