Preparing for 2024 edition: manual changes
Running `cargo fix --edition` in previous commit produced a lot of
output. Vast majority of it is about drop order of temporaries changing
in 2024.
I briefly looked through some of those warnings, but I did not see
anything interesting. I doubt we rely on this drop order anywhere.

The other type of warning is about the RPIT lifetime capture rules.
There is a case that the tool does not handle automatically and only
warns about:
https://doc.rust-lang.org/edition-guide/rust-2024/rpit-lifetime-capture.html#migrating-cases-involving-apit
I went through those warnings, and performed the changes manually.