CachingSession: Mostly fix a race condition
A race can happen when multiple calls with different statements (none of
which are in cache) are executed at the same time, with cache full.

It is possible that all of them will remove the same entry from the
cache, and then add their own entries. Then size of the cache will
exceed self.max_capacity, so we will never remove from cache again.

First fix that I thought of is to change "==" to "<=". With that fix, we
would still be able to remove elements from the map after race happens.
Note that we only remove an element when inserting a new element, so
this would still never decrease the cache overflow unless 2 cache misses
with the same query happen at the same time. So the memory leak would
still happen, but much slower.

So in addition to that I changed  "if cache full: remove" to "while
cache full: remove". In some cases it could lead to some queries getting
evicted for no reason, hurting performance. There is also slim
possibility of some thread getting starved for a bit.
Those issues seem not very probable, so may it is better to risk them
than the memory leak? I'm honestly not sure.

I also thought of checking cache size in the happy path, but len() is
not a simple operation in dashmap - it goes through shards and sums
their lengths. I didn't benchmark it, but I'm afraid it could hurt
performance.

Maybe in the future we should research how dedicated cache libraries are
approaching this?