Use precise capturing where possible
Rust 2024 introduced precise capturing. Automatic migration tools helped
with moving almost all places to use it, but some patterns are not
supported by this tool.
This commit aims to use precise capturing wherever possible.
Interestingly, I had to move one function to the top level, because it
was not working otherwise, and was giving a very unexpected error. I
don't blame Rust for not handling it well - it was a function inside a
function inside a lambda inside a function. We should burn in hell for
creating such abomination and remove other such things.