PreparedStatement: put id and lwt in shared state
PreparedStatementSharedData was initially introduced for performance
reasons, specifically to make cloning PreparedStatement cheaper.
Now that we have it, I think we should put all shared state there. If we
don't:
- It is confusing. If I see shared state, and I see that e.g. `id` is
not a part of it, then I automatically begin to wonder if its not
shared, of if its somehow mutable.
- (Minor) More stuff in shared -> cheaper cloning. Difference, if any,
should be tiny.

prepare_tracing_ids unfortunately can't be moved because it is pub.