book.yml: Set default toolchain globally
setup-rust-toolchain sets installed toolchain as default only for
current directory. `mdbook test` creates a dir in /tmp and executes
`rustdoc` there, leading to issues due to different toolchains being
used. Until it is fixed in the action or in mdbook, we need to manually
set stable toolchain as global default.