PreparedStatement: Introduce mutable metadata
This introduces second ResultMetadata field to PreparedStatement:
current_result_metadata. It is stored in ArcSwap, and updateable via new
`update_current_result_metadata` field.
The original field (`initial_result_metadata`) is kept. It can't be
removed because of backwards compatibility. The only way it is exposed
is public `get_result_set_col_specs` method, which will become
deprecated in future commit. Internal getter (`get_result_metadata`) is
renamed to `get_current_result_metadata`, and code in connection.rs
adjusted. Because of how the code is structured and all the preparation
in previous PRs, this is enough to cover all cases (_iter, retries etc).

Currently the update method is unused, intentionally. It will become
used when we implement Scylla's metadata extension.