protocol features: Add field for metadata id
For now we don't parse the feature, or mark it as supported. I want to
keep the commits atomic and tests passing. If I mmarked this feature as
supported, Scylla would send and expect the additional fields / flags in
EXECUTE / Result:Prepared / RESULT:Rows, which would cause tests to
fail.