pager: fix panic when runtime is shut down
Historically, when making changes to the code we would forget to send
something to the Pager Worker channel before closing it. This would
cause the `recv().unwrap()` in QueryPager to panic, because the channel
would be closed without having sent anything.
To harden against such logic bugs, we added an abstraction of
PageSendAttemptedProof, which enforces that at least one item is sent
to the channel before it is closed, or else `worker.work()` can't
return. With this abstraction in place, we long believed the `unwrap()`
to be safe.

However, there is one more case when `recv()` can return None: when
the runtime is being shut down. In that case, the task polling the
channel is cancelled, and the channel is closed without sending
anything.

This commit handles this case by returning an empty page when `recv()`
returns None. This allows for graceful handling of runtime shutdown
scenarios, without panicking.
We are sure that we won't introduce silent errors this way, because
if we get None, the only possible explanation is that the runtime
is indeed being shut down. The logic bugs on the side of the Pager
Worker are already handled by the PageSendAttemptedProof abstraction.

Alternatively, we could return a new NextPageError variant indicating
that the runtime is being shut down, but that seems like an overkill -
this variant would be very special and pollute the API. Returning
an empty page should be sufficient, as if someone is trying to shut down
the runtime, they likely don't care about the results anyway.

Fixes: #1435