scylla-cql: Response/Result with deser metadata
We want to deserialize metadata of Result:Rows eagerly in `scylla`. The
issue is that `scylla` operates on `Response` type, which is an enum
that stores raw metadata. We can't change definition of this enum
because of backwards compatibility. The solution here is to introduce
another enum, ResponseWithDeserializedMetadata, and allow conversion of
Response to it.