integration: introduce SerializeValueWithFakeType
This helper struct allows serializing a value with a different
CQL type than its actual Rust type. This is useful for testing
scenarios where we want to force serializing a value to a CQL column
of a non-matching type (or, more specifically, a type that is not
accepted by the type-checking logic in the driver).

In the next commit, this will be used to insert tuples with fewer
elements than the CQL type expects, to later verify that reading
such tuples does not cause deserialization errors.

Co-authored-by: Karol Baryła <karol.baryla@scylladb.com>