Connection: Update PreparedStatement metadata
In the protocol there is a new flag, "metadata_changed". In scylla-cql,
it is present on RawMetadataAndRawRows, but not on
DeserializedMetadataAndRawRows. We don't need to put this flag there and
check it explicitly. Id on result metadata is present if either this is
result metadata from PREPARED message, or the flag was present on
RESULT:Rows. This means we can check if ID is present and different from
current one.