pager: add timeouts to SingleConnectionPagerWorker
The SingleConnectionPagerWorker now supports request timeouts
when fetching pages. This is for consistency with the main
PagerWorker implementation.

Note that for now, no metadata queries specify timeouts, so
the default per-Session execution profile's timeout will be used.
Nothing prevents us from specifying custom timeouts for metadata
queries in the future, thanks to this commit.