Fix test_schema_await_with_unreachable_node
In some test cases we expect schema agreement to fail, for example when
we prevent schema fetch query to a coordinator for DDL.
In this case it is possible that schema doesn't propagate to some node,
and we issue DROP KEYSPACE to such node, resulting in an error.
Awaiting schema agreement after clearing proxy rules should fix that.