Schema agreement: Precise error handling logic
This changes the logic of schema agreement to allow some error to end
the process immediately, without waiting until the timeout.
For now the error clasification is not done with a lot of effort, just
to have something that doesn't treat transient errors as non-transient,
but also classify some errors as non-transient.