openssl: Set connect state
It informs openssl that the configuration is for a tls client, not tls
server. It is not really necessary for us, because it is clear what
openssl method that we call when connecting that it is a tls client, but
I thought it will be a bit more future proof.

For more info see:
https://docs.openssl.org/master/man3/SSL_set_connect_state/#notes