openssl: Set connect state
It informs openssl that the configuration is for a tls client, not tls
server. It is not really necessary for us, because the correct state is
deduced from us calling `connect` on `SslStream`, but I thought it will
be a bit more future proof and secure.

For more info see:
https://docs.openssl.org/master/man3/SSL_set_connect_state/#notes