openssl: Enable hostname verification
This assumes that nodes presents a certificate with its ip address in
SAN list. This is the same assumption we have for rustls.
The only alternatives I can think of is to have a hardcoded domain name
accepted for all nodes, or some callback mapping node ip to expected
SAN. Since this is not possible for rustls (you specify domain/ip only
when conmnecting, not on ClientConfigBuilder, and we can't extend
TlsContext::Rustls023 variant), I decided against introducing tests for
this or documenting this, and went with the assumption that certs are
for node IPs.