custom_type_parser: fully-qualify call to collect_array
Recently, a `collect_array` method has been introduced to the standard
library's `Iterator` trait. This method complains with `Itertool`'s
method of the same name, and the (nightly) rust compiler warns about the
future incompatibility.

Address the potential future issue by fully-qualifying the current call
to `collect_array` and leaving a comment to rewrite it to use the
standard library version of this method after it gets stabilized.

See rust-lang/rust#149266 to track the progress of stabilization.