cluster: notify HostListener about topology changes
Upon each topology refresh, the HostListener is notified about all nodes
that were added or removed from the cluster. This is done by comparing
the previous and the new `known_nodes`. When a node is added, the
listener is first notified about the addition, and then about the node
being up. When a node is removed, the listener is first notified about
the node being down (if it was up), and then about the removal. This is
in line with CPP Driver's behaviour.