{ser,de}/value: ignore frozen flag in type check
Due to a bug introduced when unifying ColumnType with legacy CqlType in
5fe4f79a23d130ca2bdf663c055888c7f7aaaa35, the `frozen` flag was
inappropriately required to be set to `false` during deserialization and
serialization type checks for collections. This caused failures when
deserializing or serializing collections with manually crafted
ColumnType instances that had `frozen` set to `true`, even though the
actual data was compatible.

The fix is to ignore the `frozen` flag during type checks for
collections, allowing deserialization and serialization to proceed
as long as the collection and element types match, regardless
of the `frozen` status.