cluster/state: add topology change handler param
Constructing `ClusterState` is the proper place to handle topology
changes. As the responsibility of handling topology changes is on
ClusterWorker, I wanted to keep this logic there instead of introducing
it to `ClusterState::new`. Therefore, I decided the following:
This commit adds a parameter to the `ClusterState::new`, which accepts
a closure from ClusterWorker that will be called with old and new
`known_peers` upon each construction of `ClusterState`.

For now, this closure is a no-op, but the next commit will implement
the actual logic of handling the topology changes in ClusterWorker.