makefile,CI: enable all features in tests
Our tests haven't been running with all features explicitly enabled.
This meant that some features could not be properly tested, because
they were not enabled in the test runs.
What made this come unnoticed is the combination of two factors:
- the "full-serialization" feature was enabled, which enables a lot of
  functionality, but not all of it;
- some features were enabled because of feature unification in Cargo,
  i.e., if any dependency enables a feature, it is enabled for the whole
  build. This meant that some features were enabled in tests, but not
  because of our explicit configuration, but because of **examples**.

This is fixed by enabling all features explicitly in test runs in CI
and Makefile.