integration: skip tests for Cassandra in code
The normal way to skip tests that are not supposed to run with Cassandra
is to use the `#[cfg_attr(cassandra_tests, ignore)]` attribute on the
test function. However, there were two tests that were being skipped
in the Cassandra CI workflow instead. This commit modifies those tests
to use the attribute instead, ensuring consistency in how tests are
skipped across the codebase. This cleans up the invocation in CI.