codewide: add scylla_unstable flag
This flag is going to mimic `tokio_unstable` flags.
It's goal is to allow enabling unstable features in the driver,
like experimental APIs, or APIs that are not fully ready yet,
or unsafe / type-unsafe APIs that require extra caution and are not
intended for regular users.

As of now, such features include:
- unstable-testing
- cpp_rust_unstable
- nodejs_rs_unstable

So in the further commits I'll make those features be only elegible
when this flag is set.

The flag is opt-in, so by default the driver remains stable.
However, as the only thing that this flag does is to allow enabling
unstable features, it seems OK to enable it in CI and Makefile,
so that we can test the unstable features as well.
I added a special Makefile target to test without the flag as well,
just in case, and in CI I added a step to cargo check without the flag.