Cargo.toml: reorder items for better grouping
I'm tired of seeing then in unintuitive order.
Now they are grouped:
- package metadata
- lints
- benchmarks
- features
- dependencies

Dependencies go last, because it's a long list that we rarely need
to peruse. Conversely, package metadata and lints are both short and
more important to see at a glance.

Also, benchmarks are now next to features, because they depend on
features.