Pass trough reconnect policy
This commit makes progress towards making the ReconnectPolicy
configurable. Before, it was hardcoded in PoolRefiller config.
It is not yet present in SessionConfig, it will be done in next commit.
For now, the default is hardcoded in two places: Session::connect, and
PoolConfig::default (this is only for tests).

I decided to put ReconnectPolicy in PoolConfig, but not put
ReconnectPolicySession in HostPoolConfig. Why?
ReconnectPolicy is, imo quite obviously, element of connection pool
configuration. ReconnectPolicySession however is mutable - it performs
work. For that reason it doesn't fit as just element of configuration.
To handle that, I changed the conversion from PoolConfig to
HostPoolConfig to return a pair of HostPoolConfig and
ReconnectPolicySession.