Add jitter to ExponentialReconnectPolicy
It works the same way as in cpp-driver. I added it here instead of
cpp-rs-driver, because I think it will be a useful mechanism for Rust
Driver too. PoolRefiller already has a problem of opening all
connections in a fill simultaneously. If there are multiple clients, it
can result in a big connection storm. Adding a jitter should lessen the
impact at least a bit.