Remove connection pool from MetadataReader
I see 3 possible advantages of connection pool:
 1. It uses mpsc for error notifications, instead of oneshot. This makes
worker code simpler.
 2. It automatically reconnects when connection is lost.
 3. Creating it is not async, it creates connections in the background.

(1) can be worked around as this commit proves.
(2) is not really used by us. When receiving error notification, we
immediately refresh metadata, which should almost always happen before
pool reconnect succeeds, so the pool will be dropped.
(3) is not a problem. MetadataReader constructor is already async
because of contact point resolving.

The big disadvantage of the current design (with the pool) is that it
makes it impossible to order CQL events relative to CC drops and
reconnects.
This prevents implementing features (like event channels, or partial
metadata fetches) in a robust way, but also makes code more difficult to
understand.

The other disadvantage is that pool returns a `Connection`. This forces
us to create `ControlConnection` on each fetch, instead of when
connection is created. This in turn makes it more difficult to make
metadata reading use prepared statements (by caching them in
`ControlConnection`). I also think any contributor will agree that
storing `ControlConnection` instead of `Connection` or `ConnectionPool`
in `MetadataReader` makes more conceptual sense.

Side-effect of this change is that control connection will no longer be
included in metrics about connections. This will affect, for example,
tests in cpp-driver. Interestingly, this matches the behavior of legacy
cpp-driver. If there is demand for bringing those metrics back, we could
do that without using a connection pool.

This commit removes this use of `ConnectionPool`. Instead, `Connection`
is now used.